feat: :sparkles: add section for "what is kendo?"
What I am adding to the home page on this branch is a combination of excerpts/edits of content on our old kendo website and content authored by me. Credit for the content from the previous website goes to the people who wrote it (there is no record of who is credited for writing what, so I am unable to specify that exactly).