feat: :fire: remove social media links after considering values and possible alternate locations
Had a value discussion considering why the social media links would be on the website and what value they'd bring to the club. Best solution: Relocate the Facebook group link to another location more directly associated with members as opposed to the general public. Instagram doesn't need to be linked for now.