fix: :bug: make calendar render every time
full-calendar.ts was being imported every time, but the code inside it was only running the time it was being initialized (the first time it was imported) because that's how the module system in js works. In order to fix that, we put the code to create and render the calendar in a function and exported it, and we have it run after every import.

Co-authored-by: Nate Stringham <nstringham@users.noreply.github.com>