feat: :sparkles: improve clarity and consistency of writing
Made names for parts of the shinai used in paragraphs lowercase to be consistent and because they are not proper nouns. Made parts of the shinai section more concise by removing unnecessary phrases in the descriptions of parts. Reworded certain sentences to be in active voice instead of passive. Removed references to locations of photos on the page to account for responsive design.