fix: :bug: use correct netlify env var, make URL public, rename it
Use public env variable instead of private to reflect actual way we are using it (browser does have access to it because we send it as a reponse to their request). Use DEPLOY_PRIME_URL from netlify bc it changes based off of particular deploy instead of just being prod base url.

Co-authored-by: Nate Stringham <nstringham@users.noreply.github.com>